Functions for clean and -j flag
